requirements.txt
google-generativeai>=0.3.0
python-docx>=0.8.11
pyinstaller>=6.0.0
customtkinter>=5.2.0
pillow>=9.0.0
gradio>=4.0.0 